move configuration reading logic into common, logic now shared between connector and cli
