+++ crates/common/src/lib.rs
@@ -3,4 +3,5 @@ pub mod clickhouse_parser;
 pub mod client;
 pub mod config;
 pub mod config_file;
+pub mod format;
 pub mod schema;
